Attempt extraction of Resume document content in Prawn block into another file
