+++ lib/resume_generator/cli/application.rb
@@ -48,7 +48,7 @@ module ResumeGenerator
         require 'prawn'
         require 'prawn/table'
         inform_start_of_resume_generation
-        Resume::Document.generate(self)
+        Resume::Generator.start(self)
         inform_of_successful_resume_generation
       end
