+++ lib/resume_generator/cli/file_system.rb
@@ -4,11 +4,11 @@ module ResumeGenerator
       def self.open_document(app)
         case RUBY_PLATFORM
         when %r(darwin)
-          system("open #{Resume::Document.filename}")
+          system("open #{Resume::Generator.filename}")
         when %r(linux)
-          system("xdg-open #{Resume::Document.filename}")
+          system("xdg-open #{Resume::Generator.filename}")
         when %r(windows)
-          system("cmd /c \"start #{Resume::Document.filename}\"")
+          system("cmd /c \"start #{Resume::Generator.filename}\"")
         else
           app.request_user_to_open_document
         end
